move OnPaintFinished and DisposeTask to skiapaint class
